McuManager MTU; ImageManager sets mtu based on err
This commit modifies McuManager, ImageManager, FirmwareUpgradeManager,
and McuMgrBleTransport to be more like the Android library when handling
MTU and image upload. In otherwords, All managers start with a default
MTU which may be changed by the developer or if the transport returns
and insufficientM error.

The default MTU calculation has also been modified to select the MTU
based first on scheme (ble vs udp) and then iOS version:

 * BLE:
     - iOS 11+: 524
     - iOS 10: 182
     - iOS 9: 155
 * UDP: 1024

This commit also moves some of the transport agnostic errors that were
originally in McuMgrBleTransport.swift McuMgrError to
McuMgrTransport.swift McuMgrTransportError. The BLE specific errors are
still in McuMgrBleTransport.swift as McuMgrBleTransportError.