Added 'image' parameter to ImageManager (#59)
As part of a change in DFU specification, a new 'image' parameter was added as part of the payload to support devices with multiple cores and their slots. If this is approved, I'll try to follow-up up with a change to FirmwareUpgradeManager to allow easier access to this new capability.