Multi-Image Support in FirmwareUpgradeManager (testMode-Only)
This is a comprehensive change that aims to maintain backwards-compatibility with single-image upload as before, as well as with the new multi-image support. Big changes were required to the logic, though in principle, the previous code structure and system is maintained, if only patched to allow the code to be more aware of the complete process.

.testAndConfirm is implemented, however, the firmware forces the mode to .testOnly upon detection of Multi-Image since the newest available firmware will not support Confirming, because at the moment swap-back of images is not possible inside the updated device.