Pipeline now has Concurrency Enabled
However, it's only being used with a depth of '1', which means it's effectively serial guarded by the Semaphore. And also, we can't 'read' the value of the Semaphore ourselves, so there's not a lot of point to keeping the depth ourselves. I thought we could print it, but it's a shame that the wait() call doesn't return the value.