Split McuMgrBleTransport's internal lock into two
The part that we need to work concurrently is the write calls, not the rest. The rest makes sense for it to happen only once at a time. So, the first step to figure this out is to separate one thing from the other. Let's see if we can make this happen.