McuMgrBleTransport Step towards multi-packet write-lock
We needed a way to allow multiple packets to be sent at the same time. And whilst this has not been achieved here yet, we made a step, which is that we can store multiple locks now per packet, and we can track the packets by reading their sequence number. And if we can't read the sequence number, we throw an error. This might need to be walked back later, because changing the API itself was a big mess, but I think this looks good. We will see.