Changes in McuMgrBleTransport ahead of switching to returning a Result<>
I was not aware of the importance of the logic within the fail() and success() private functions, that was my issue. I'm slowly trying to untangle this, and to begin with I've limited the use of connectionLock so that it doesn't stop potential concurrent calls from doing their own work.