McuMgrBleTransport's _send() now returns a Result and no longer takes in a callback
The tricky part here was the parsing of the Data into the Generic Response Type because if we did this inside the _send() function, the typing got lost up the chain and when passing it down into the DispatchQueue.main for the callback. So we had to move the parsing over there for this to work. The API send() function is now more convoluted, but overall I think this is saner code, at least the internal _send().