Switch back to single 'writeLock' in McuMgrBleTransport
If we're only going to do a single callback for multiple data packets being sent, we don't need to track multiple writes / sequence number if we only want to do a single 'return' so to speak. We still need to do the same work, or update the logic, for the response side of things, but for now we can commit this bit of work at least.