Backtrack: Restored previous McuManager API and McuMgrBleTransport no longer emits multiple writes
So, the problem here is that our pipeline speed was limited by our strategy of writing 'in blocks', rather than a pipeline / window scheme wherein there's always 'pipelineDepth' write requests in flight. So, we made McuMgrBleTransport not aware of pipeling anymore, but aware or resilient towards multiple writes, and now we need to thrust the logic on the upper levels.
The upside is that we're not breaking the McuManager API, which in my opinion is important for current users of this library.