Cleanup: New McuMgrBleTransportWriteState class
So a lot of mutable state protection logic was in the 'write' or 'send' logic code, and we've moved it out into its own class. This allows us to use a serial DispatchQueue as a lock to protect writes and reads of the shared state, and although it's cleaned up a lot of things, it looks like it might be slightly slower / more CPU hungry. I'll check with the previous commit.