Speed Improvement to McuMgrBleTransportWriteState
So now it's a Dictionary, rather than an array. We had so many calls to find out the index of an item, and I thought a lot of CPU power was being spent in constantly checking an index, so it seems like a Dictionary here is the correct data structure. I'm not doing proper Instruments measurements, but I think CPU usage is lower by at least 4% if not at times 6% or 10%, so I'm going to call this an improvement and leave it at that.