In 'Basic' Image Upload Mode, a valid Byte Alignment Setting is now required
In 'Advanced', we expect the user to know what they're doing. But in Basic, we want to make it easier for the user to get things done. And Pipelining requires a Byte Alignment to be set or otherwise we cannot predict the offsets the firmware is going to take, that's the basic pillar of being able to 'jump' offsets faster with more data chunks in place. So this change ensures there's a valid Byte Alignment setting if the user decides to use Pipelining.