More Code Changes ahead of SMP Reassembly
The Buffer Size is an extra parameter we need to move around, but why do we need another parameter when we already have a Data Structure for this? So we made upstream changes in the API to get the proper Data.
Also, there's code in the Transport to 'chunk' the Data down at the Transport level if it's larger than the MTU, which you might remember, was our first 'pipelining breakthrough'. Though I didn't look back at that code.