Custom 1-second timeout for McuMgrParameters Request
There's a bug in NCS-1.9 firmware that makes it not reply with the expected RC 8 Error. So, we need to assume the bug won't be fixed for the vast majority of devices, so what we do instead is timeout very fast, after 1 second.
This means we needed to extend / change the API to take in a custom timeout parameter now.