Fix for 'missed offset' Pipeline Stall
This is a patch, not a proper logic fix. But it's basically that if the expected offsets don't match, we're left with duplicate response offsets, the pipeline writes don't continue, and the upload stalls because our logic believes it's waiting for a response, but it won't, because of the missed offset. So the fix is, if we receive an offset we're not expecting, we clear out the lowest value in the pipeline, and the rest of the logic sorts itself out (maybe?). I saw this codepath being called and it was because the upload was restarted with a different offset, so it does do something and work, but whether it's reliable I don't know yet. But it's better.