"Pipeline Depth" is now "Number of Buffers"
So, code-wise the code configuration value is still 'pipeline depth', because that's the better term for what the code does. For now. But for the UI, Filip asked for the feature to get renamed towards the user. And the relationship is that the number of buffers is one plus the pipeline depth, so we had to fix that, and update the labels properly, and fix the Storyboard segues and UIActions and stuff because Xcode doesn't know how to rename things. But anyway, I think we're there.