Fix for failed McuMgrParameters Request if Device is not already Connected
We were sending the McuMgrParameters request without checking if we're connected to the device first. This is not an issue if you always connect to the Device, such as sending 'Hello', and then attempting a DFU. But if all you do is connect and go straight to DFU, then the request is missed (sent before it can be replied to) and therefore the speed increase is lot. Now we just wait for validateCallbac(), which is guaranteed to get called once we're connected to the Device.