Fix for Confirmation-Loop Bug
Bug was caused by starting a muilti-image DFU Upload wherein the images were in the wrong order. The for-loop inside confirmationCallback expects self.image[j] to match j, otherwise we might mark the incorrect image as confirmed but skip over the one that is not confirmed in the loop itself (guard statement checking our own properties whether an image is confirmed). This extends further into test and validate stages, where we also made some changes to the code.