Fix for 'Missed ACK' Log Messages
The root cause were Responses being received out-of-order than what we expected. The fix was to track which is the offset we expect for each packet sent, represented by the Sequence Number. But instead of doing it all in a single commit, we went the multiple-commit approach for a better insight into the changes.