Fix: Retry instead of Automatic Failure for Timed-Out .write(s)
Previously, if during Upload a package was not delivered, it'd stall the UI for up to 30 seconds (standard timeout) and then fail. Now, we detect the 'write timeout' case and retry with a much shorter duration, allowing for cases wherein a couple of missed writes / packages don't cause the entire process to fail.