Fix for 3x "Slow Timeouts" Failure
We want timeouts to be retried when they fail fast, not when they take a long time. If the device takes 30s to reply back, it's likely it won't reply after a further 60 seconds, so McuMgrBleTransport has added logic to differentiate this with a new 'fast timeout' threshold set to 10 seconds.