Unified Sequence Number(s): McuManager now tracks Sequence Numbers and rotates them for every single command/write/send
Previously, only ImageManager was handling sequence numbers, and that too, only for upload. All other issued commands / writes defaulted to sequence number zero. I think a possible cause for issues in the validated / timeout stage was that when issuing commands that were not the upload, we were going back to sequence number zero. And they might've been dropped / confused by the rest of the code. Now they're integrated within McuManager, so they're used for all commands. This required deeper changes into the ImageManager logic, but I think we might've solved a couple of previous issues.

Maybe.