Fix for McuMgrBleTransport not appending Data from 'chunked' Response
Subsequent chunks of Data belonging to the same SequenceNumber are not pre-pended with the header, which includes the Sequence Number. This means that if a response is 'chunked', the subsequent chunks will be parsed for a header and most likely return an invalid Sequence Number value, therefore the WriteState will not append that Data to the proper Sequence Number, therefore never returning the Data to the upper levels.