Fix: Erase App Settings does not Fail Fast Enough
Some firmware does not even reply back to the 'Erase App Settings' command, which was not being taken into account. So now we do, and if we get that specific error, we stop attempting to Erase App Settings directly and try to complete the DFU process as quickly as possible.