Fix: If Device Capabilities are not Supported, then Erase App Settings Isn't
There's no point in sending an Erase App Settings Command if the firmware device is not going to reply back to it, which in turn leads to the Upload DFU to fail when it doesn't need to.
Unfortunately at the moment we do not turn back the switch in the UI, but we do log the issue.