Added 'Erase App Settings' as a FirmwareUpgradeState value
Without it, it was not visible exactly what the app / DFU was doing when the Erase App Settings caused the issue in older versions of firmware. That has been addressed of course, but still, improving visibility of what the app is doing is a good thing for now and the future.