Fix: BasicManager's echo() returns an Error if the message is too long
There's a limit to the size of the echo message, which is 2475 bytes, package size (not payload - tested). If the message is over this limit, the firmware device will just not reply - wish the behaviour was different, but it's not. So, we provide a helpful error message that returns quickly rather than waiting for a timeout Error.