Fix: When uploading offset 0 of any Image, we need to wait more before triggering a Timeout
This is because writing a new Image might trigger an erase on the firmware side, which could take a few seconds. Also, note that even if we've set '20' seconds as the timeout number, the total timeout time will probably be closer to 60, because the BleTransport has 3 automatic retries unless an error is encountered.