Fix for invalid lock on Connection when sending a Packet
McuMgrBleTransport sets a connection lock with a timeout of 20 seconds, but then when it's time to wait for said lock, it waits for just a second, which might not be enough, causing a timeout, but then, the next time a packet is attempted (due to the min number of retries being larger than 1), the connection might be set because a bit more time was needed. This might mean the Mcu Mgr Parameters Request might fail just because the target firmware device was not super fast or BLE traffic was high.