Fix for bad default Filesystem Mount Point
This was changed in Zephyr / McuMgr SDK to /lfs1 from the previous /lfs. User can edit this in Example app by tapping on 'Edit' on the upper-right corner, but they shouldn't need to do this to get this feature to work with a fresh DK from one of our own samples. Error Code has also been updated to better inform users.