Initial Direct XIP Support + Revert of Default Image Slot Value (back to 0)
This commit represents the first time we were able to upload Direct XIP without having to stop the debugger, check something, modify something in the code, etc. Now, this doesn't mean that it's finished - I believe it can be made a lot tidier. But it's closer.

As for the revert on the Default Image Slot value, if we leave Direct XIP aside, what we want every time is for the image to be in slot 0. That's what our comparison for 'is this already uploaded?' is against. Hence, why the correct value, I think now after going back and forth a couple of times, is zero for how we've been doing things up until now.