Minor Changes
Just trying to familiarise myself with the code again after a recent break, and trying to tighten up some bits of code a bit more. Although I can't think of an alternative to the embedded for in LIST to mark the other Slot for the same Image as Uploaded, which is executed twice. Other than checking that's not marked as Uploaded, of course.