Fix for Confirm Loop
This fixes two Confirm Loops, actually. The code was not written expecting having to send a 'Confirm' to Slot zero, so that caused issues. Then, when that was addressed, there was an infinite loop, but now because after uploaded to 0-1, 0-0, with 0.0 Active, if we try to confirm 0-1, it doesn't return that it has been Confirmed in the response, sending us into an infinite loop. So we mark it as confirmed after we send the command. If there is an error, anyway, the callback should catch it and return the Error.