DirectXIP Confirm Loop Fix
A version of this was written, but commented out for 1.4's release since 'standard behaviour' dictates that firmware should not behave like the sample code we have that does and that causes this issue. Regardless, we think it's a good thing to deal with it because of how important DFU is.