Added status, crc32, sha256 and closeAll APIs to FileSystemManager
These were missing from the previous release of iOS-McuMgrLibrary.
Also, interesting wrinkle here: none of this is tested :( So, if you're swearing at this code via git-blame yes, blame me. I wrote a full frontend to implement and test Pipelining for FS Upload, but I'm not doing that for these.