API Change: 'mode' from FirmwareUpgradeManager is now set through FirmwareUpgradeConfiguration
I think this makes more sense. I wish I'd made this change originally, but corrections & long-time improvements can happen later. And more changes & fixes that we're working on for nRFCDM will show 1.4 was kind of rushed, and it was. But it was also important for us to meet that deadline and, we live in a real world. We met the deadline, it was not perfect, but we did it. We can keep improving things now.