DirectXIP Upload is now Successful if either Slot Hash Matches
So basically, we won't upload the inactive Slot and try to run from it like we would before. If either slot 0 or 1 has a matching hash to our Image, in DirectXIP (which has 2 hashes for 2 slots), we declare it successful. The way we implemented this is not the most elegant way, but it works. That's the important part.