BLE Transport Layer Re-Order Buffer Improvements
The ROB (Re-Order Buffer) now effectively returns back values when it's not missing the intervening numbers. This should speed-up DFU speed in situations where sequence numbers are returned out of order. It could be argued this is not necessary and that the need for re-ordering responses is indeed a red-herring for other DFU Library issues, which is why the Buffer can now log. So if the ROB Buffer is indeed reordering responses, they should be visible on the logs, hence providing information about a potential cause for an issue.