SUIT (Software Update for Internet of Things) Support
This breaks API compatibility, so it requires build to be bumped up to 1.5. Since we need to release this ASAP, and there's no time for me to properly test writing against this API in nRF Connect, we will release the app but not release library updates. This will wait, and presumably be released as 1.5.1 if changes are required.
I tried to add ".suit" file specifically, but it does not work. So I had to open up the DocumentController to allow any file to be selected, which we can correct later.