Streamlining of start() API for McuMgrSuitEnvelope
The hash part is a bit complicated, so we've made the API better from the library user's perspective. We were on the fence regarding making the modes 'public' and so on, but these things can change specially since SUIT is under heavy development. So if we need to, we can open it up later.