SUIT Bootloader Introduction + Internal-use only McuMgrSuitManifest
There's more functionality that needs to be built upon parsing the bootloader's name. This is just a start. And as for the McuMgrSuitManifest, we'd really like to implement a full parser, but this is as far as we got before we had to move on. So for now, it's not marked as Public.