Change: Don't set Bootloader to "mcuboot" if it's unknown
Before, we set it to "suit" if we didn't know. Now, it's better to only decide if a "file" is a SUIT file if the content matches specifically the envelope. If there's an MCU Boot Version set, it's more likely it's an MCU Boot image, but if not, let's not make any assumptions.