McuMgrPackage now parses a SUIT Manifest .ZIP
So, for SUIT Zip Packages, we do not care about the other files. We only look for the SUIT Envelope in the Manifest, and that file has a hash, or a digest we can sink our teeth onto, and then we can proceed with SUIT. But again, this is just one small step.