Cleanup step: No more 'suitMode' in FirmwareUpgradeConfiguration
This was an "implementation detail" leak, mostly because we had no view into whether we were doing SUIT or not. We have more information now, and understand SUIT better. Most importantly, through the existing API we can specifically track whether the update is for SUIT or not if needed through other means that are not just the bootloader, so we are able to cleanup our API.