Fix: McuMgrPackage no longer considers itself the owner of your Caches Directory
This I think, is not an issue on iOS. But it is on macOS, where the Caches Directory returned by the API is not Application-specific, but shared. So if we clean it out and then unzip there, we're essentially deleting stuff we don't own. Also, like it happened in nRF Connect, we might be deleting the same file we want to DFU !