FirmwareUpgradeController API Expansion
So, if you're a user of the API, you need a way, when you get the SUIT Delegate Callback, to provide the Data. Most API users, we assume, will keep their reference to the Upgrade Manager of their own, as is the case of the Example App. But, if you don't and you might be using the Delegate's API's returned protocol FirmwareUpgradeController, you might need a way to provide your resource there. So if we expanded the Controller's API to support this. I also had the thought of killing this 'Controller' protocol since it's mostly the FirmwareUpgradeManager itself. But, I thought giving us more API flexibility in the future might be a good idea.