SuitListResponse
For SuitManager's list, we now return a subtype of McuMgrResponse. This keeps the API in line with all the others. As for why this was not the case originally, it's because this API is not simple. We need to send one call, and build the response out of other responses. So there's no 'one McuMgrResponse' that we can return, so we build it instead. Hopefully we built it well.