Added CustomStringConvertible & Codable Conformance to multiple types
This is to make it easier to surface information in the UI after making a (SUIT) List Request. Which I kind of don't know yet exactly what it does nor what it informs us about - but I hope that when I do, we don't have to modify the API too much.