Optimization: Reduced Memory Pressure by autoreleasing generated Strings in hexEncodedString()
We flagged the _send() function in McuMgrBleTransport as being responsible for increase in memory pressure when performing DFU. And the culprit, was appending Strings. We don't do enough complicated Strings in the _send() function itself, except, that we log all the bytes that we send. And we send a lot of bytes in a very small amount of time, so this optimization makes sense here.