Added Support for SUIT 'Cache' Image Upload & Deferred Install(s)
The procedure has now been slightly altered. If Image(s) with .suitCache content is detected for upload, we will send the "defer_install" flag, finish uploading the Envelope, proceed uploading all Cache Images, send the equivalent "Confirm" command, and then poll in case any additional resources are requested. At the moment, polling is not supported by existing firmware, but it may be in the future.