Fix for Duplicate Payload Building Implementation(s)
There are two independent implementations for calculating a packet's payload, one for the packet itself, and the other to calculate the packet header's size for each offset, to allow us to calculate then how much Data we can put in to fit within our MTU bounds. This meant that every time we need to alter the payload, we'd have to remember to modify this twice. And it has not happened once, nor twice, that I've missed the second place where the payload is calculated. It's time to put a remedy to this ongoing situation.