Chance: McuMgrPackage Image(s) now includes 'caches'
This is because cache(s), or SUIT Cache(s), are actually Image(s) that also need to be uploaded. Furthermore, in the case where the McuBoot Bootloader is used to Upload SUIT, this maps beautifully as well considering they target different partitions or 'images'. So it's both a code and a logic  / design win.