Fix: Length of Advanced Commands Text is now refreshed on Update
Before this change, the cell containing the text of the 'Read' / 'List' Command for example would grow in size, but never update when the displayed text shrank, so you could end up with blank space everywhere except in the middle of the screen. This is a brute force approach, but oh well.