Fix to allow Partition Selection in Advanced Mode only for SUIT
We need to allow, in 'Advanced' Mode, for the user to select multiple .bin files and do the update themselves. This means, we want to allow the user to select to which Partition they'd like to send the .bin file. We hacked this earlier, making a single .bin destined for SUIT to show up as multiple 'Image's. Which is... a hack. So we decided to put the hacky code somewhere else that makes more sense, which is in the UI. This has led to other API changes, primarily the code to get the 'name' of an Image.