"Barcode" Issue Improvement
In Advanced Mode, with a File Selected and if we perform a Read / List command and the output is fairly long, when switching between tabs the UI might look like a stretched barcode at times. This is due to the blunt instrument that is 'reloadData' applied on the UITableView. This is a fractional improvement.