Bumped Write Buffer Size Constant from '10' to '16'
So, what is the write buffer size? We know the underlying CoreBluetooth API has a buffer to store all the Data we 'give it' to send to a CBPeripheral, but we don't know how big it is. If we ask CBPeripheral if it's ready to receive new Data, it almost always doesn't end up well. Perhaps we could do some further work to look into it. I might in fact, look into that next. But, regardless, increasing this number should hopefully prevent us from "waiting" and entering the spinlock scenario of waiting to hear back from CBPeripheral.