Change: McuMgrBleTransportWriteState now has an open() API for Error(s)
We can convert this API call into a 'sync' instead of 'async'. But in general, we should prefer 'async' calls instead of 'sync'. There's a reason why we went the 'sync' reason here which is, I presume, if there's an error we want to stop immediately and not do unnecessary work. We will do more testing to confirm whether this breaks anything or not.