Change: Limit Reassembly Buffer Size to fall within CoreBluetooth API Buffer
The way we understand it at the moment, CoreBluetooth has an underlying buffer that accumulates all the Data passed to it via cbPeripheral "write" API calls. If we exceed this buffer, the Data passed-in to it is just lost. When we think we might hit this limit, we stop and ask the CBPeripheral if we think we can start writing again. Problem is, this almost always happens, as we have the code written now, in the middle of a buffer write. And the stall is for so long, that the firmware starts thinking we wrote the chunk of Data already. It's a mess. There are levers in the current API to fix this, like reducing the number of buffers, reassembly size, etc. But we want to keep the library / API more reliable, hence the nature of this change.