Fix for Incorrect Upgrade Mode Override for 'Upload Only'
This override is designed to make SUIT over MCUBoot work seamlessly, but it turns out we're just overriding the Upgrade Mode all the time if it's ever set to 'Upload Only'. So now, we only override the Upgrade Mode if we detect that the content that we're sending is SUIT over MCUBoot Bootloader.