Rebuilt BLE Transport Layer
The previous "ROBBuffer", was never a ROB Buffer as such, since it did no reordering work. Instead, it tracked Out-of-Order buffer response callbacks, so we renamed it to reflect that. And instead, we wrote a new component below the McuMgrBleTransport, that does reordering of write operations, and henceforth does receive the "ROBWriteBuffer" name. The purpose of this change is to try to improve how we try to recover from "Peripheral is not ready for write without response" situations, or at the very least, not make things worse by retrying sequence numbers and losing the coherence of the packets we're sending. All packets we send, must follow their sequence number, so we cannot send chunk 0, 1 of Seq. 25, and then switch to Seq. 24 and send chunk 0 before finishing all chunks of Seq. 25 - it'll just throw off the firmware. And when we retry a previous sequence number, we can do so, but not "break" our current sequence of pending writes. This is... complicated.