Added Internal Soft Reset for McuMgrBleTransport
Customers of the API will probably, like nRF Connect, build a new transport for each DFU attempt. But it's possible others do like nRF Connect Device Manager, and keep reusing the same Transport. It's possible this affects the repeated DFU attempts with unclean state of private / internal properties. So we're going to try to get better at detecting new connections, and using those as an opportunity to reset internal variables that might need it.