Independent McuMgrTransport for BaseViewController (Possible fix for hitting assert on McuMgrBleTransportWriteState's newWrite(sequenceNumber:lock:) functionj)
BaseViewController sets up the McuMgrTransport instance for all subclasses to use. Unfortunately, the transport layer, specifically McuMgrBleTransport, is getting more and more complicated. And its code runs under the assumption that it contains a single "user". It is perfectly fine for multiple managers to share a transport, but they cannot be performing requests at the same time (in parallel). Because if they do, like BaseViewController does whenever one of its child VCs causes the shared CBPeripheral's connection state to change to .connected, it's going to fire its own set of requests. And, those requests might fly through on the same McuMgrBleTransport, or, they could land on the exact same McuSequenceNumber that the other one is using, and therefore hitting the assert that expects all new write operations, for a given sequence number, to either have an open lock or not exist.

Translation: BaseViewController and FirmwareUpgradeManager's requests for Bootloader Information were happening in parallel and using the same sequence number ID. So when whoever got there second tried to send its request, a 'lock' for said Sequence Number already existed and was awaiting a response.

Or at least, that's what I think is happening.