Fix: Override Peripheral Is Not Ready For Write Without Response
Since we have measured the cutouts, we've never seen more than about 6ms. The previous, like, three releases ago, version waited for about 15ms. That seemed to be a good ballpark; so what we're doing, is ignoring it. Filip does the same, and it doesn't appear to get Mesh app hurt. But Filip's strategy is the reverse, which is to send something, and wait for "peripheralIsReady()" callback. That's an option we could pursue, flipping the logic around. Potentially less speed, I think. But we can keep that idea on the back of our minds.

Anway, I cannot properly test this change because... CoreBluetooth. Or Bluetooth. I don't... anymore.