Attempted Support for Group-Specific Errors in FileSystemManager
Some errors from the Filesystem Group, such as errors when writing to a file, etc. need to be read correctly via their Group RC code, instead of defaulting to the standard McuMgr Error Group. Otherwise, the cause of a failure might not be understood correctly. Like, "Error Code 8" means "Busy, Try Later" in Standard but is actually "Write Failed" for Filesystem.

"Attempted" because I could replicate the error last night, but lost the code in a Git mishap, and now I can't replicate it.