Fix for Incorrect Transport Setup on Filesystem McuMgrParameters Response
If one part of the stack sends Data as if we're doing Reassembly but the lower one isn't aware, it's not going to work. The same code structure works on FirmwareUpgradeManager because it uses the McuMgrUploadPipeline, which handles Reassembly. We can use it here as well, but let's go step by step.