Fix: Delay Pipelining (though not implemented library-side currently) and Reassembly after first successful Sequence Number Upload for FileSystemManager
If upload is going to fail, for example because there's no space to store the file we're going to send, let's fail "faster". Better feedback for the user. This is also reminiscent of what the Android library does overall. I think this will make FileSystem more reliable. We could also port this behaviour back to FirmwareUpgradeManager. But one by one. Let's fix FileSystemManager first.