Fix in FileSystemManager for "No Reply" to Upload Packet from SMP Server
We need to cap the MTU size at 495 bytes. If the Data packet is larger than 495 bytes, it will not get replied to / acknowledged by SMP Server firmware. By setting the MTU to 495, we're sending less than 495, but for the 4 bytes we're missing, it's not worth making the code / comments more difficult to understand. 495 is the number, after all.