Patched McuMgrBleTransport's initial MTU Value
This should be rare, in the sense that peripheral.maximumWriteLength(for: .withoutResponse) should provide a good number that's correct. But we have feedback, like here https://github.com/NordicSemiconductor/IOS-nRF-Connect-Device-Manager/issues/361 that we suspect, a value of '20' must be wrong. And anyway, any operation could never go through with a value of 20 because it's smaller than the header size, so nothing is going to work. So we're basically patching it to always be the smallest value it can be. And also yes, modified a bit more of code since it was around.