Preliminary (as in, we'd be surprised if there are no issues) support for Pipelining in FileSystemManager
So, it was possible to enable pipelining via FirmwareUpgradeConfiguration 'configuration' argument. But it was not implemented, so no speed improvement or changing in behaviour would occur. Now, it is enabled internally via the same McuMgrUploadPipeline, shared with ImageManager and SuitManager. Preliminary tests show it working as expected, but FileSystemManager is complicated to test due to the small external flash storage availability. But we're here to solve issues.